// src/api/generated/endpoints/account-api.ts
export interface AccountApiChangePasswordRequest {
    /**
     * 
     * @type {ChangePasswordDto}
     * @memberof AccountApiChangePassword
     */
    readonly changePasswordDto?: ChangePasswordDto
}

/**
 * Request parameters for mySubjects operation in AccountApi.
 * @export
 * @interface AccountApiMySubjectsRequest
 */
export interface AccountApiMySubjectsRequest {
    /**
     * 
     * @type {'JSS1' | 'JSS2' | 'JSS3' | 'NUR1' | 'NUR2' | 'PRY1' | 'PRY2' | 'PRY3' | 'PRY4' | 'PRY5' | 'PRY6' | 'SS1' | 'SS2' | 'SS3'}
     * @memberof AccountApiMySubjects
     */
    readonly _class?: MySubjectsClassEnum

    /**
     * 
     * @type {number}
     * @memberof AccountApiMySubjects
     */
    readonly pageNumber?: number

    /**
     * 
     * @type {number}
     * @memberof AccountApiMySubjects
     */
    readonly pageSize?: number

    /**
     * 
     * @type {string}
     * @memberof AccountApiMySubjects
     */
    readonly search?: string
}

/**
 * AccountApi - object-oriented interface
 * @export
 * @class AccountApi
 * @extends {BaseAPI}
 */
export class AccountApi extends BaseAPI {
    /**
     * 
     * @param {AccountApiChangePasswordRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AccountApi
     */
    public changePassword(requestParameters: AccountApiChangePasswordRequest = {}, options?: RawAxiosRequestConfig) {
        return AccountApiFp(this.configuration).changePassword(requestParameters.changePasswordDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AccountApi
     */
    public getMyProfile(options?: RawAxiosRequestConfig) {
        return AccountApiFp(this.configuration).getMyProfile(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {AccountApiMySubjectsRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof AccountApi
     */
    public mySubjects(requestParameters: AccountApiMySubjectsRequest = {}, options?: RawAxiosRequestConfig) {
        return AccountApiFp(this.configuration).mySubjects(requestParameters._class, requestParameters.pageNumber, requestParameters.pageSize, requestParameters.search, options).then((request) => request(this.axios, this.basePath));
    }
}

/**
  * @export
  * @enum {string}
  */
export enum MySubjectsClassEnum {
    JSS1 = 'JSS1',
    JSS2 = 'JSS2',
    JSS3 = 'JSS3',
    NUR1 = 'NUR1',
    NUR2 = 'NUR2',
    PRY1 = 'PRY1',
    PRY2 = 'PRY2',
    PRY3 = 'PRY3',
    PRY4 = 'PRY4',
    PRY5 = 'PRY5',
    PRY6 = 'PRY6',
    SS1 = 'SS1',
    SS2 = 'SS2',
    SS3 = 'SS3'
}